Changes from a Create and delete action to a ListUpdate logic
Depending on the Add/Remove action of the Lists(ListofProjecty,InvoiceLists,Expenses) there are the necessari action taken to keep the DB in sync.

Todo currently the List gets changed and after the change the i "Try" to sync to DB if this fails its Async here i need to make sure that the app and the DB stay in sync!